Add user authorization for credits services

+++ src/main/java/tn/esprit/gestionbancaire/services/impl/AdministrativeDocumentServiceImpl.java
@@ -3,15 +3,16 @@ package tn.esprit.gestionbancaire.services.impl;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import tn.esprit.gestionbancaire.enums.CreditStatus;
 import tn.esprit.gestionbancaire.exception.ErrorCodes;
 import tn.esprit.gestionbancaire.exception.InvalidEntityException;
 import tn.esprit.gestionbancaire.exception.InvalidOperationException;
 import tn.esprit.gestionbancaire.model.AdministrativeDocument;
 import tn.esprit.gestionbancaire.model.Credit;
+import tn.esprit.gestionbancaire.model.User;
 import tn.esprit.gestionbancaire.repository.AdministrativeDocumentRepository;
 import tn.esprit.gestionbancaire.services.AdministrativeDocumentService;
 import tn.esprit.gestionbancaire.services.CreditService;
+import tn.esprit.gestionbancaire.utils.Utility;
 import tn.esprit.gestionbancaire.validator.AdministrativeDocumentValidator;
 
 import java.time.Instant;
@@ -54,7 +55,9 @@ public class AdministrativeDocumentServiceImpl implements AdministrativeDocument
 
     @Override
     public AdministrativeDocument update(AdministrativeDocument administrativeDocument) {
+
         Credit credit = creditService.findById(administrativeDocument.getCredit().getId());
+        checkAuthorization(credit.getUser());
         List<String> notes = credit.getNotes();
         notes.add(administrativeDocument.getAdministrativeDocumentType() + "Document has been updated");
         credit.setNotes(notes);
@@ -64,17 +67,25 @@ public class AdministrativeDocumentServiceImpl implements AdministrativeDocument
 
     @Override
     public List<AdministrativeDocument> findAllByCreditId(Integer id) {
+        Credit credit = creditService.findById(id);
+        checkAuthorization(credit.getUser());
         return administrativeDocumentRepository.findAllByCreditId(id);
     }
 
     @Override
     public AdministrativeDocument findById(Integer id) {
+        Optional<AdministrativeDocument> administrativeDocument = administrativeDocumentRepository.findById(id);
+
+        if (administrativeDocument.isPresent()) {
+            checkAuthorization(administrativeDocument.get().getCredit().getUser());
+        }
         return administrativeDocumentRepository.getById(id);
     }
 
 
     @Override
     public void delete(Integer id) {
+        isAdmin();
         if (id == null) {
             log.error("Administrative Document ID is null");
             return;
@@ -82,6 +93,7 @@ public class AdministrativeDocumentServiceImpl implements AdministrativeDocument
         Optional<AdministrativeDocument> administrativeDocument = administrativeDocumentRepository.findById(id);
 
         if (administrativeDocument.isPresent()) {
+            checkAuthorization(administrativeDocument.get().getCredit().getUser());
             administrativeDocument.get().setCredit(null);
             administrativeDocumentRepository.save(administrativeDocument.get());
             administrativeDocumentRepository.deleteById(id);
@@ -98,6 +110,7 @@ public class AdministrativeDocumentServiceImpl implements AdministrativeDocument
         Optional<AdministrativeDocument> byId = administrativeDocumentRepository.findById(id);
         AdministrativeDocument administrativeDocument = null;
         if (byId.isPresent()) {
+            checkAuthorization(byId.get().getCredit().getUser());
             administrativeDocument = byId.get();
             administrativeDocument.setPhoto(imageURL);
             administrativeDocumentRepository.save(administrativeDocument);
@@ -108,5 +121,22 @@ public class AdministrativeDocumentServiceImpl implements AdministrativeDocument
             administrativeDocument.setLastModifiedDate(Instant.now());
         }
         return administrativeDocument;
+    }
+
+    private void isAdmin() {
+        User currentUser = Utility.getCurrenUser().getUser();
+        if(!currentUser.getRoles().equals("admin")) {
+            throw new InvalidOperationException("UNAUTHORIZED USER MUST BE ADMIN",
+                    ErrorCodes.UNAUTHORIZED);
+        }
+    }
+
+    private void checkAuthorization(User user){
+        User currentUser = Utility.getCurrenUser().getUser();
+        if ((!user.getId().equals(currentUser.getId())) && !currentUser.getRoles().equals("admin") ){
+            throw new InvalidOperationException("UNAUTHORIZED USER MUST BE ADMIN OR CREDIT OWNER",
+                    ErrorCodes.UNAUTHORIZED);
+        }
+
     }
 }
